Drop the casts the relaxed operand rule made dead
A uint answers `__index__` and now meets a bare integer as well, so a range,
a slice, a subscript and a comparison all read it directly. That retires 136
casts in the tests and 35 in the specs, and lets a few spec helpers say what
they mean: a registry index advances with `validator_index + 1`, a builder
index masks with `validator_index & ~BUILDER_INDEX_FLAG`.

What stays is what the library still forces. A repeat keeps its cast, because
`Uint64.__rmul__` raises where returning NotImplemented would let CPython fall
back to sequence repetition. A yielded value keeps its cast, because YAML has
no representer for a uint. Sibling unit types keep theirs, which is right: a
slot times an epoch has no unit anyone would want.

+++ tests/core/pyspec/eth_consensus_specs/test/fulu/unittests/das/test_das.py
@@ -33,12 +33,12 @@ def test_compute_matrix(spec):
     blob_count = 2
     input_blobs = [get_sample_blob(spec, rng=rng) for _ in range(blob_count)]
     matrix = spec.compute_matrix(input_blobs)
-    assert len(matrix) == int(spec.CELLS_PER_EXT_BLOB * blob_count)
+    assert len(matrix) == spec.CELLS_PER_EXT_BLOB * blob_count
 
     rows = chunks(matrix, spec.CELLS_PER_EXT_BLOB)
     assert len(rows) == blob_count
     for row in rows:
-        assert len(row) == int(spec.CELLS_PER_EXT_BLOB)
+        assert len(row) == spec.CELLS_PER_EXT_BLOB
 
     for blob_index, row in enumerate(rows):
         # The first half of a row's cells holds the original (un-extended) blob data
